feat(api): added socket and orgnaized routes and socket files in proper way
