+++ apps/api/package.json
@@ -15,8 +15,13 @@
     "license": "ISC",
     "dependencies": {
         "@tsconfig/node21": "^21.0.1",
+        "add": "^2.0.6",
         "cors": "^2.8.5",
-        "express": "^4.18.2"
+        "express": "^4.18.2",
+        "fs": "0.0.1-security",
+        "https": "^1.0.0",
+        "pnpm": "^8.15.3",
+        "socket.io": "^4.7.4"
     },
     "devDependencies": {
         "@types/cors": "^2.8.17",
