+++ apps/web/package.json
@@ -31,6 +31,7 @@
         "react-day-picker": "^8.10.0",
         "react-dom": "^18.2.0",
         "react-router-dom": "^6.22.0",
+        "socket.io": "^4.7.4",
         "tailwind-merge": "^2.2.1",
         "tailwindcss-animate": "^1.0.7",
         "vaul": "^0.9.0"
